app/requirements.txt
Pillow
numpy
fastapi
uvicorn
python-multipart
git+https://github.com/WHOIGit/provenance-capture.git
